gdansk-league/scripts/requirements.txt
# Python dependencies for Riot API data collection

requests==2.31.0
python-dotenv==1.0.0
supabase==2.3.4
pandas==2.1.4
